requirements.txt
flask
flask-cors
requests
python-dotenv
g4f
typing-extensions
curl_cffi
markdown
